place mediawiki under supervisord
We're changing over to using supervisord for daemon control, because we also
want to deploy a microservice which exposes update.php to a http endpoint.

This will allow us to ask the cron service to execute runjobs, instead of
requiring a cronjob on the docker host.

This makes semawi significantly more portable and "correct".