speed up loading of the semawi data model
If 50+ minutes load time is speeding up...

We've downgraded mysql; version 8 was taking twice as long, and I see a lot of
stories in github and on mediawiki of strange issues. So 5.7 looks to be the one
to stick on for now.

We're also splitting up dump importing and link rebuilding. It still takes a
long time but there's results in the actual wiki a lot faster.