make the walled garden capability opt-out
Since we're going to be running iotwiki.dk also, that wiki has the walled garden
settings disabled. Rather than operate with manually edited LocalSettings.php
files, it's preferable to split that off as a tunable in `env-sample` and
`.env`.

That way both semawi and iotwiki can share the same codebase, and only the local
.env will differ (as intended).